# docs/api/openapi.json
{
  "components": {
    "schemas": {
      "Body_append_scan_image_scans__scan_id__images_post": {
        "properties": {
          "image": {
            "contentMediaType": "application/octet-stream",
            "title": "Image",
            "type": "string"
          },
          "image_quality": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Image Quality"
          },
          "image_sha256": {
            "title": "Image Sha256",
            "type": "string"
          },
          "panel": {
            "title": "Panel",
            "type": "string"
          }
        },
        "required": [
          "panel",
          "image_sha256",
          "image"
        ],
        "title": "Body_append_scan_image_scans__scan_id__images_post",
        "type": "object"
      },
      "Body_create_deferred_scan_scans_deferred_post": {
        "properties": {
          "buyer_type": {
            "default": "RETAIL",
            "title": "Buyer Type",
            "type": "string"
          },
          "captured_at": {
            "title": "Captured At",
            "type": "string"
          },
          "category": {
            "title": "Category",
            "type": "string"
          },
          "client_uuid": {
            "title": "Client Uuid",
            "type": "string"
          },
          "coverage_asserted": {
            "title": "Coverage Asserted",
            "type": "boolean"
          },
          "dimensions": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Dimensions"
          },
          "ecommerce": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ecommerce"
          },
          "flags": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Flags"
          },
          "geo": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Geo"
          },
          "investigation_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Investigation Id"
          },
          "mode": {
            "title": "Mode",
            "type": "string"
          },
          "package_shape": {
            "default": "RECTANGULAR",
            "title": "Package Shape",
            "type": "string"
          },
          "panels": {
            "items": {
              "type": "string"
            },
            "title": "Panels",
            "type": "array"
          },
          "scale_reference": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Scale Reference"
          }
        },
        "required": [
          "client_uuid",
          "captured_at",
          "mode",
          "category",
          "coverage_asserted",
          "panels"
        ],
        "title": "Body_create_deferred_scan_scans_deferred_post",
        "type": "object"
      },
      "Body_create_scan_scans_post": {
        "properties": {
          "buyer_type": {
            "default": "RETAIL",
            "title": "Buyer Type",
            "type": "string"
          },
          "captured_at": {
            "title": "Captured At",
            "type": "string"
          },
          "category": {
            "title": "Category",
            "type": "string"
          },
          "client_uuid": {
            "title": "Client Uuid",
            "type": "string"
          },
          "coverage_asserted": {
            "title": "Coverage Asserted",
            "type": "boolean"
          },
          "dimensions": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Dimensions"
          },
          "ecommerce": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ecommerce"
          },
          "flags": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Flags"
          },
          "geo": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Geo"
          },
          "image_quality": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Image Quality"
          },
          "image_sha256": {
            "items": {
              "type": "string"
            },
            "title": "Image Sha256",
            "type": "array"
          },
          "images": {
            "items": {
              "contentMediaType": "application/octet-stream",
              "type": "string"
            },
            "title": "Images",
            "type": "array"
          },
          "investigation_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Investigation Id"
          },
          "mode": {
            "title": "Mode",
            "type": "string"
          },
          "package_shape": {
            "default": "RECTANGULAR",
            "title": "Package Shape",
            "type": "string"
          },
          "panels": {
            "items": {
              "type": "string"
            },
            "title": "Panels",
            "type": "array"
          },
          "scale_reference": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Scale Reference"
          }
        },
        "required": [
          "client_uuid",
          "captured_at",
          "mode",
          "category",
          "coverage_asserted",
          "panels",
          "images",
          "image_sha256"
        ],
        "title": "Body_create_scan_scans_post",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "title": "LMPC Compliance API",
    "version": "1.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/scans": {
      "get": {
        "operationId": "list_scans_scans_get",
        "parameters": [
          {
            "in": "query",
            "name": "manufacturer",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Manufacturer"
            }
          },
          {
            "in": "query",
            "name": "brand",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Brand"
            }
          },
          {
            "in": "query",
            "name": "category",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Category"
            }
          },
          {
            "in": "query",
            "name": "status",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          },
          {
            "in": "query",
            "name": "overall",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Overall"
            }
          },
          {
            "in": "query",
            "name": "date_from",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "date",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Date From"
            }
          },
          {
            "in": "query",
            "name": "date_to",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "date",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Date To"
            }
          },
          {
            "in": "query",
            "name": "officer_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "uuid",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Officer Id"
            }
          },
          {
            "in": "query",
            "name": "investigation_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "uuid",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Investigation Id"
            }
          },
          {
            "in": "query",
            "name": "jurisdiction_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "uuid",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Jurisdiction Id"
            }
          },
          {
            "in": "query",
            "name": "violation_type",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Violation Type"
            }
          },
          {
            "in": "query",
            "name": "q",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 200,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Q"
            }
          },
          {
            "in": "query",
            "name": "page",
            "required": false,
            "schema": {
              "default": 1,
              "minimum": 1,
              "title": "Page",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "page_size",
            "required": false,
            "schema": {
              "default": 25,
              "maximum": 100,
              "minimum": 1,
              "title": "Page Size",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "sort",
            "required": false,
            "schema": {
              "default": "newest",
              "title": "Sort",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response List Scans Scans Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "List Scans",
        "tags": [
          "review"
        ]
      },
      "post": {
        "operationId": "create_scan_scans_post",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_create_scan_scans_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Create Scan",
        "tags": [
          "scans"
        ]
      }
    },
    "/scans/deferred": {
      "post": {
        "description": "Declare a field inspection without images.\n\nA repeated client_uuid returns the existing scan, so the retry that timed out\non a rural network never creates a second inspection.",
        "operationId": "create_deferred_scan_scans_deferred_post",
        "requestBody": {
          "content": {
            "application/x-www-form-urlencoded": {
              "schema": {
                "$ref": "#/components/schemas/Body_create_deferred_scan_scans_deferred_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Create Deferred Scan",
        "tags": [
          "scans"
        ]
      }
    },
...
    "/scans/{scan_id}/complete-upload": {
      "post": {
        "description": "Confirm that every declared panel image has arrived.",
        "operationId": "complete_scan_upload_scans__scan_id__complete_upload_post",
        "parameters": [
          {
            "in": "path",
            "name": "scan_id",
            "required": true,
            "schema": {
              "title": "Scan Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Complete Scan Upload",
        "tags": [
          "scans"
        ]
      }
    },
...
    "/scans/{scan_id}/images": {
      "post": {
        "description": "Upload one declared panel image idempotently.",
        "operationId": "append_scan_image_scans__scan_id__images_post",
        "parameters": [
          {
            "in": "path",
            "name": "scan_id",
            "required": true,
            "schema": {
              "title": "Scan Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_append_scan_image_scans__scan_id__images_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Append Scan Image",
        "tags": [
          "scans"
        ]
      }
    },
...
  }
}